replace "FakeSMTP" class with factory class + targeted stubbing
That way the tests are even closer to the actual (production) code as
we just stub out interaction with real "socket" instances.
(We could actually run pymta listening on a local socket but that would
make the tests much slower and more brittle as the there is a lot of
peculiarities of the underlying operating system.)

The real motivation driving this change was the desire to test
upcoming SMTP logging. I need to add some important logging hooks
in the ".send()"/".getreply()" methods and "FakeSMTP" used these
to intercept the communication.
Now these methods are untouched in testing code so we can also test
interaction happening there.