split default log files for CLI tools into "delivery log" and "queue log"
The logged messages of actual deliveries and queued emails will be different
in the future. Automated parsing of log files in a uniform format is likely
easier.

I also hope this helps with investigating problems manually, as it will be
immediately obvious whether a message was queued or delivered via SMTP.